added user

+++ processes/mqtt-consumer.js
@@ -2,7 +2,7 @@ const Parse = require('parse/node');
 const { mqttClient } = require('../mqtt-client');
 const { defineSessionToken, createLog } = require('../functions');
 
-const { gate, gateTransponder } = global.gConfig.topicParts;
+const { main, gate, gateTransponder } = global.gConfig.topicParts;
 const { checkin, checkout, charge, ok } = global.gConfig.messageParts;
 let sessionToken;
 
@@ -11,7 +11,7 @@ async function start() {
     try { sessionToken = await defineSessionToken() }
     catch (err) { console.error(err) }
 
-    mqttClient.subscribe(`+/${gate}/${gateTransponder}`, {}, (err, ok) => {
+    mqttClient.subscribe(`${main}/${gate}/${gateTransponder}`, { qos: 1 }, (err, ok) => {
         if (err) {
             console.error("[MQTT] subscription", err.message);
         };
@@ -45,6 +45,7 @@ function getRoom(roomNumber) {
 }
 
 function getRoomTowel(roomNumber) {
+    console.log(roomNumber)
     const innerQuery = new Parse.Query('Room')
         .equalTo('num', roomNumber)
 
@@ -75,15 +76,22 @@ function handleCheckin(roomNumber) {
         });
 }
 
-function handleCheckout(roomNumber) {
+async function handleCheckout(roomNumber) {
+    getRoom(roomNumber)
+        .then(room => {
+            room.set('isOccupied', false);
+            room.set('lastCheckOut', new Date())
+            return room.save(null, { sessionToken })
+        })
+        .then(() => { createLog(`room ${roomNumber} was checked out of`) })
+        .catch((err) => {
+            console.error(err.message)
+            createLog(`failed checking out of room ${roomNumber}, ${err.message}`)
+        });
     getRoomTowel(roomNumber)
         .then(roomTowel => {
-            createLog(`guest ${roomTowel.get('guestName')} has checked out of room ${roomNumber}`)
-            return roomTowel.destroy({ sessionToken });
-        }).catch((err) => {
-            console.error(err)
-            createLog(`failed cheking guest ${roomTowel.get('guestName')} out of room ${roomNumber}`)
-        })
+            if (roomTowel) { return roomTowel.destroy({ sessionToken }); }
+        }).catch((err) => { console.error(err) })
 }
 
 function setStatusRefunded(refund) {
@@ -94,16 +102,15 @@ function setStatusRefunded(refund) {
 
 async function handleCharge(serial, answer) {
     const transaction = await getPendingTransaction(serial);
-    answer == ok
-        ? transaction.set('status', transaction.get('action') == 'refund'
-            ? 'approved'
-            : 'ok'
-        )
-        : transaction.set('status', 'denied');
-
-    if (transaction.get('refund')) {
-        try { await setStatusRefunded(transaction.get('refund')) }
-        catch { err => { } }
+    if (answer == ok) {
+        transaction.set('status', 'ok')
+        if (transaction.get('refund')) {
+            try { await setStatusRefunded(transaction.get('refund')) }
+            catch { err => { } }
+        }
+    }
+    else {
+        transaction.set('status', 'denied');
     }
 
     transaction.save(null, { sessionToken })
